Refactor page token proto.
