+++ src/main/kotlin/org/wfanet/measurement/kingdom/service/api/v2alpha/BUILD.bazel
@@ -116,6 +116,7 @@ kt_jvm_library(
         "//src/main/proto/wfa/measurement/api/v2alpha:event_group_kt_jvm_proto",
         "//src/main/proto/wfa/measurement/api/v2alpha:event_groups_service_kt_jvm_grpc",
         "//src/main/proto/wfa/measurement/api/v2alpha:event_groups_service_kt_jvm_proto",
+        "//src/main/proto/wfa/measurement/api/v2alpha:page_token_kt_jvm_proto",
         "//src/main/proto/wfa/measurement/internal/kingdom:event_group_kt_jvm_proto",
         "//src/main/proto/wfa/measurement/internal/kingdom:event_groups_service_kt_jvm_grpc",
         "//src/main/proto/wfa/measurement/internal/kingdom:event_groups_service_kt_jvm_proto",
@@ -227,6 +228,7 @@ kt_jvm_library(
         "//src/main/proto/wfa/measurement/api/v2alpha:measurement_spec_kt_jvm_proto",
         "//src/main/proto/wfa/measurement/api/v2alpha:measurements_service_kt_jvm_grpc",
         "//src/main/proto/wfa/measurement/api/v2alpha:measurements_service_kt_jvm_proto",
+        "//src/main/proto/wfa/measurement/api/v2alpha:page_token_kt_jvm_proto",
         "//src/main/proto/wfa/measurement/api/v2alpha:protocol_config_kt_jvm_proto",
         "//src/main/proto/wfa/measurement/internal/kingdom:measurement_kt_jvm_proto",
         "//src/main/proto/wfa/measurement/internal/kingdom:measurements_service_kt_jvm_grpc",
